+++ src/service2/rooms.util.ts
